refactor the news component
