+++ packages/kilo-gateway/src/tui.ts
@@ -25,8 +25,7 @@ export { registerKiloCommands } from "./tui/commands/kilo-commands.js"
 export { DialogKiloTeamSelect } from "./tui/components/dialog-kilo-team-select.js"
 export { DialogKiloOrganization } from "./tui/components/dialog-kilo-organization.js"
 export { KiloAutoMethod } from "./tui/components/dialog-kilo-auto-method.js"
-export { NotificationBanner } from "./tui/components/notification-banner.js"
-export { DialogKiloNotifications } from "./tui/components/dialog-kilo-notifications.js"
+export { KiloNews } from "./tui/components/kilo-news.js"
 
 // ============================================================================
 // Re-exported Types
