+++ packages/kilo-gateway/src/tui/components/dialog-kilo-notifications.tsx
@@ -31,7 +31,7 @@ export function DialogKiloNotifications(props: DialogKiloNotificationsProps) {
     <box paddingLeft={2} paddingRight={2} gap={1}>
       <box flexDirection="row" justifyContent="space-between">
         <text attributes={TextAttributes.BOLD} fg={theme.text}>
-          Notifications
+          News
         </text>
         <text fg={theme.textMuted}>esc</text>
       </box>
